Moved timestamp classes to separate timestamp.py.
